requirements.txt
streamlit
langchain-huggingface
langchain-chroma
openai
sentence-transformers
pysqlite3-binary
tiktoken
langchain-qdrant 
qdrant-client